fix(pullsecret): remove pullSecret feature from chart
The CI is currently broken because KinD cannot pull the images as it is stored in
`docker.pkg.github.com` and therefore is private. We need that `pullSecret` feature to be able to
pull the images. New images will be pushed on `ghcr.io` soon and will not need the `pullSecret`
feature anymore.