feat(timed): add static file volume
The timed backend (especially the admin section) needs some static files
to work.  Given Django best-practices, it is not recommended for the
static files to be served by Django itself, but rather by a dedicated
web server.

This creates a volume for the static files, which can then be shared with
the frontend web server, which already provides the application assets.

The static files are generated by the startup script of the backend
container, so they will be freshly written any time the conainer comes
up. We don't need to worry about them being stale etc.