feat: only create service with serviceMonitor
Only create the service when `serviceMonitor.create` is set. The service
is useless otherwise. I have a project where I can only use the
`prometheusRules.enabled=true` but only want to use the prometheus
rules, not the service monitor. Also, it creates a conflict in our
deployment because we already have a similar service with that name
`vault-monitoring`.